refactor(step4): split component, extract row + helpers, RGAA polish (#3355)
- Extract `step4/quartileErrors.ts` (FieldErrorMap, deriveErrors, buildRecap, hasAnyError, fieldId)
- Extract `step4/quartileFormHelpers.ts` (toQuartileData, normalizeForMutation, gipToQuartiles, emptyQuartiles, computeMinsForTable)
- Extract `step4/QuartileTableRow.tsx` (one row of the inverted quartile table) for readability
- Step4QuartileDistribution.tsx down from 572 → 358 lines (under 400-line cap)
- Use `~/modules/shared` barrel for `useZodForm`
- RGAA 11.10: append "en euros" to threshold input aria-label so screen readers know the unit (the visible "€" stays aria-hidden)
