+++ packages/app/src/modules/declaration-remuneration/steps/step4/QuartileTable.tsx
@@ -1,12 +1,9 @@
 "use client";
 
-import { QUARTILE_NAMES } from "~/modules/declaration-remuneration/shared/constants";
-import type {
-	QuartileData,
-	QuartileTuple,
-} from "~/modules/declaration-remuneration/types";
-import { computePercentage, displayDecimal } from "~/modules/domain";
+import type { QuartileTuple } from "~/modules/declaration-remuneration/types";
+import { computePercentage } from "~/modules/domain";
 import stepStyles from "../Step4QuartileDistribution.module.scss";
+import { QuartileTableRow } from "./QuartileTableRow";
 
 type Field = "threshold" | "women" | "men";
 
@@ -33,54 +30,6 @@ type Props = {
 	onQuartileChange: (index: number, field: Field, value: string) => void;
 	disabled?: boolean;
 };
-
-/** Format ordinal text like "1er quartile" → 1<sup>er</sup> quartile */
-function OrdinalLabel({ text }: { text: string }) {
-	const match = text.match(/^(\d+)(er|e)\s(.+)$/);
-	if (!match) return <>{text}</>;
-	const [, num, suffix, rest] = match;
-	return (
-		<>
-			{num}
-			<sup>{suffix}</sup> {rest}
-		</>
-	);
-}
-
-function fieldId(
-	tableType: "annual" | "hourly",
-	index: number,
-	suffix: "max" | "f" | "h",
-) {
-	return `step4-${tableType}-q${index + 1}-${suffix}`;
-}
-
-function getQuartileSuffix(index: number) {
-	return index === 0 ? "er" : "e";
-}
-
-const TYPE_LABEL: Record<"annual" | "hourly", string> = {
-	annual: "annuel",
-	hourly: "horaire",
-};
-
-function thresholdAriaLabel(tableType: "annual" | "hourly", index: number) {
-	return `Seuil maximum ${index + 1}${getQuartileSuffix(index)} quartile ${TYPE_LABEL[tableType]}`;
-}
-
-function womenAriaLabel(tableType: "annual" | "hourly", index: number) {
-	return `Nombre de femmes ${index + 1}${getQuartileSuffix(index)} quartile ${TYPE_LABEL[tableType]}`;
-}
-
-function menAriaLabel(tableType: "annual" | "hourly", index: number) {
-	return `Nombre d'hommes ${index + 1}${getQuartileSuffix(index)} quartile ${TYPE_LABEL[tableType]}`;
-}
-
-function pct(q: QuartileData, gender: "women" | "men") {
-	const total = (q.women ?? 0) + (q.men ?? 0);
-	if (total === 0) return "-";
-	return computePercentage(q[gender] ?? 0, total);
-}
 
 export function QuartileTable({
 	title,
@@ -143,169 +92,18 @@ export function QuartileTable({
 										</tr>
 									</thead>
 									<tbody>
-										{quartiles.map((q, i) => {
-											const isLast = i === quartiles.length - 1;
-											const idMax = fieldId(tableType, i, "max");
-											const idF = fieldId(tableType, i, "f");
-											const idH = fieldId(tableType, i, "h");
-											const cellErrors = errors[i];
-											const thresholdErr = cellErrors?.threshold;
-											const womenErr = cellErrors?.women;
-											const menErr = cellErrors?.men;
-											return (
-												<tr key={QUARTILE_NAMES[i]}>
-													<th scope="row">
-														<OrdinalLabel text={QUARTILE_NAMES[i] ?? ""} />
-													</th>
-													<td className={stepStyles.minCell}>{mins[i]}</td>
-													<td>
-														{isLast ? (
-															<span className={stepStyles.readonlyCell}>
-																- €
-															</span>
-														) : (
-															<div
-																className={
-																	thresholdErr
-																		? `fr-input-group fr-input-group--error ${stepStyles.cellInputGroup}`
-																		: stepStyles.cellInputGroup
-																}
-															>
-																<span
-																	className={`fr-sr-only ${stepStyles.cellLabel}`}
-																	id={`${idMax}-label`}
-																>
-																	{thresholdAriaLabel(tableType, i)}
-																</span>
-																<span className={stepStyles.inputWithUnit}>
-																	<input
-																		aria-describedby={
-																			thresholdErr
-																				? `${idMax}-error`
-																				: undefined
-																		}
-																		aria-invalid={
-																			thresholdErr ? "true" : undefined
-																		}
-																		aria-labelledby={`${idMax}-label`}
-																		className="fr-input"
-																		disabled={disabled}
-																		id={idMax}
-																		inputMode="decimal"
-																		onChange={(e) =>
-																			onQuartileChange(
-																				i,
-																				"threshold",
-																				e.target.value,
-																			)
-																		}
-																		type="text"
-																		value={displayDecimal(q.threshold ?? "")}
-																	/>
-																	<span aria-hidden="true">€</span>
-																</span>
-																{thresholdErr && (
-																	<p
-																		className="fr-error-text fr-mt-1v"
-																		id={`${idMax}-error`}
-																	>
-																		{thresholdErr}
-																	</p>
-																)}
-															</div>
-														)}
-													</td>
-													<td>
-														<div
-															className={
-																womenErr
-																	? `fr-input-group fr-input-group--error ${stepStyles.cellInputGroup}`
-																	: stepStyles.cellInputGroup
-															}
-														>
-															<span
-																className={`fr-sr-only ${stepStyles.cellLabel}`}
-																id={`${idF}-label`}
-															>
-																{womenAriaLabel(tableType, i)}
-															</span>
-															<input
-																aria-describedby={
-																	womenErr ? `${idF}-error` : undefined
-																}
-																aria-invalid={womenErr ? "true" : undefined}
-																aria-labelledby={`${idF}-label`}
-																className="fr-input"
-																disabled={disabled}
-																id={idF}
-																inputMode="numeric"
-																onChange={(e) =>
-																	onQuartileChange(i, "women", e.target.value)
-																}
-																pattern="[0-9]*"
-																type="text"
-																value={q.women ?? ""}
-															/>
-															{womenErr && (
-																<p
-																	className="fr-error-text fr-mt-1v"
-																	id={`${idF}-error`}
-																>
-																	{womenErr}
-																</p>
-															)}
-														</div>
-													</td>
-													<td>
-														<div
-															className={
-																menErr
-																	? `fr-input-group fr-input-group--error ${stepStyles.cellInputGroup}`
-																	: stepStyles.cellInputGroup
-															}
-														>
-															<span
-																className={`fr-sr-only ${stepStyles.cellLabel}`}
-																id={`${idH}-label`}
-															>
-																{menAriaLabel(tableType, i)}
-															</span>
-															<input
-																aria-describedby={
-																	menErr ? `${idH}-error` : undefined
-																}
-																aria-invalid={menErr ? "true" : undefined}
-																aria-labelledby={`${idH}-label`}
-																className="fr-input"
-																disabled={disabled}
-																id={idH}
-																inputMode="numeric"
-																onChange={(e) =>
-																	onQuartileChange(i, "men", e.target.value)
-																}
-																pattern="[0-9]*"
-																type="text"
-																value={q.men ?? ""}
-															/>
-															{menErr && (
-																<p
-																	className="fr-error-text fr-mt-1v"
-																	id={`${idH}-error`}
-																>
-																	{menErr}
-																</p>
-															)}
-														</div>
-													</td>
-													<td>
-														<strong>{pct(q, "women")}</strong>
-													</td>
-													<td>
-														<strong>{pct(q, "men")}</strong>
-													</td>
-												</tr>
-											);
-										})}
+										{[0, 1, 2, 3].map((i) => (
+											<QuartileTableRow
+												disabled={disabled}
+												errors={errors[i] ?? {}}
+												index={i}
+												key={`${tableType}-q${i + 1}`}
+												min={mins[i] ?? ""}
+												onQuartileChange={onQuartileChange}
+												quartile={quartiles[i] ?? { threshold: undefined }}
+												tableType={tableType}
+											/>
+										))}
 										<tr>
 											<th scope="row">Tous les salariés</th>
 											<td className={stepStyles.minCell} />
